% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_community_matrix.R
\name{extract_community_matrix}
\alias{extract_community_matrix}
\title{extract_community_matrix}
\usage{
extract_community_matrix(
  BirdPipe.data = BirdPipe.data,
  threshold = 0,
  name.to.use = "species_name"
)
}
\arguments{
\item{BirdPipe.data}{A BirdPipe data object}

\item{threshold}{Classification threshold to be used to filter the data (influences number of detected species)}

\item{name.to.use}{Which names to use for the species}
}
\value{
A community matrix, i.e. samples (=segments) times species matrix
}
\description{
Extracts a community matrix from BirdPipe data
}
\examples{
BirdPipe.data = load_data(BirdPipe.folder = "BirdPipe_demo_data");
extract_community_matrix(BirdPipe.data = BirdPipe.data);
}
