% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_validated_community_matrix.R
\name{extract_validated_community_matrix}
\alias{extract_validated_community_matrix}
\title{extract_validated_community_matrix}
\usage{
extract_validated_community_matrix(
  BirdPipe.data = BirdPipe.data,
  manual.validation.file = NULL,
  validated = c("Y")
)
}
\arguments{
\item{BirdPipe.data}{A BirdPipe data object}

\item{manual.validation.file}{Excel file of manual validation results}

\item{validated}{Which strings are considered as validated detections}
}
\value{
A community matrix, i.e. samples (=segments) times species matrix
}
\description{
Extracts a validated community matrix from BirdPipe data and manual validation file
}
\examples{
BirdPipe.data = load_data(BirdPipe.folder = "BirdPipe_demo_data", tgre);
extract_validated_community_matrix(BirdPipe.data = BirdPipe.data, manual.validation.file = "detections_validated.xlsx", validated = c("Y","YM"));
}
