% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_classification_per_segment.R
\name{highest_classification_per_segment}
\alias{highest_classification_per_segment}
\title{highest_classification_per_segment}
\usage{
highest_classification_per_segment(
  BirdPipe.data,
  threshold = 0.5,
  manual.validation.file = NULL
)
}
\arguments{
\item{BirdPipe.data}{A BirdPipe data object}

\item{threshold}{Classification threshold to be used to filter the data}

\item{manual.validation.file}{Excel file to be written for entering manual validation results}
}
\value{
A named list where the locations of highest classifications are given for each species
}
\description{
Finds the vocalization with highest classification probability for each species and segment
}
\examples{
BirdPipe.data = load_data(BirdPipe.folder = "BirdPipe_demo_data");
highest_classification_per_segment(BirdPipe.data = BirdPipe.data);
}
