% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(BirdPipe.folder, threshold = 0, geojson.folder = "geojson")
}
\arguments{
\item{BirdPipe.folder}{The name of the BirdPipe folder}

\item{threshold}{Classification threshold to be used to filter the data}

\item{geojson.folder}{The name of the geojson folder within the  BirdPipe data folder}
}
\value{
A named list with elements $recordings and $detections

A BirdPipe data object
}
\description{
Loads BirdPipe geojson data: recordings and species detections
}
\examples{
load_data(BirdPipe.folder = "BirdPipe_demo_data");
}
