% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_sound.R
\name{play_sound}
\alias{play_sound}
\title{play_sound}
\usage{
play_sound(
  BirdPipe.folder,
  audio_folder = NULL,
  audio_filename,
  offset,
  classification.interval = 3,
  buffer = 1
)
}
\arguments{
\item{BirdPipe.folder}{The BirdPipe data folder}

\item{audio_folder}{The name of the audio folder within the  BirdPipe data folder}

\item{audio_filename}{The name of the audio file}

\item{offset}{The beginning of the 3 second segment from which the vocalization is classified}

\item{classification.interval}{Length of the classification interval in seconds}

\item{buffer}{Number of seconds added before and after the classification interval}
}
\description{
Plays part of audio file (.flac or .wav) e.g. for manual validation purpose
}
\examples{
play_sound(BirdPipe.folder = "BirdPipe_demo_data", audio_filename = "field-4_2025-12-22T10_05_03.011300Z.wav", offset = 28);
#This function call plays a vocalization of Eurasian Bullfinch from the R-package demonstration data.
#Note that Mac users may obtain a "Permission denied” error. If this happens, you may need to set the folder where the audio player is located by calling e.g. tuneR::setWavPlayer("/usr/bin/afplay").
}
