% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detections.R
\name{plot_detections}
\alias{plot_detections}
\title{plot_detections}
\usage{
plot_detections(
  BirdPipe.data,
  name.to.use = "species_name",
  species.order = "alphabetical",
  jitter = 0,
  threshold = 0
)
}
\arguments{
\item{BirdPipe.data}{A BirdPipe data object}

\item{name.to.use}{Which names to use for the species}

\item{species.order}{In which order to show the species ("alphabetical" or "number_of_detections")}

\item{jitter}{the amount of random jitter added to shows overlapping datapoints}

\item{threshold}{Classification threshold to be used to filter the data (influences number of detected species)}
}
\description{
Plots BirdPipe detections as timeline
}
\examples{
BirdPipe.data = load_data(BirdPipe.folder = "BirdPipe_demo_data");
plot_detections(BirdPipe.data = BirdPipe.data)
}
