% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recordings.R
\name{plot_recordings}
\alias{plot_recordings}
\title{plot_recordings}
\usage{
plot_recordings(
  BirdPipe.data,
  show.timeline = TRUE,
  show.map = TRUE,
  background.map = NULL,
  show.number.of.species = FALSE,
  threshold = 0,
  show.segment.id = FALSE,
  number.of.species.color = "white",
  segment.id.color = "white",
  jitter = 0,
  spatial.buffer = 1
)
}
\arguments{
\item{BirdPipe.data}{A BirdPipe data object}

\item{show.timeline}{Boolean variable (TRUE/FALSE) describing whether timeline is to be plotted}

\item{show.map}{Boolean variable (TRUE/FALSE) describing whether map is to be plotted}

\item{background.map}{optional terra/SpatRaster object to be plotted as background map}

\item{show.number.of.species}{Boolean variable (TRUE/FALSE) describing whether number of detected species is shown}

\item{threshold}{Classification threshold to be used to filter the data (influences number of detected species)}

\item{show.segment.id}{Boolean variable (TRUE/FALSE) describing whether segment id:s are shown}

\item{number.of.species.color}{color for plotting number of species}

\item{segment.id.color}{color for plotting segmend id:s}

\item{jitter}{the amount of random jitter added to spatial locations to avoid their possible overlap (proportional to data extent)}

\item{spatial.buffer}{size of margins added to the map (proportional to data extent)}
}
\description{
Plots BirdPipe recordings as timeline and on a map
}
\examples{
BirdPipe.data = load_data(BirdPipe.folder = "BirdPipe_demo_data");
plot_recordings(BirdPipe.data = BirdPipe.data)
}
